backend/requirements.txt
fastapi
uvicorn[standard]
bandit
weasyprint
openai
requests
GitPython
python-dotenv